Moving dump_evaluation_metrics
Moved the func dump_evaluation_metric from mlm_training.py to multihopkg.utils_debug.dump_evals under the same function name.
